+++ modules/core/client/views/home.client.view.html
@@ -15,7 +15,6 @@
 
 
 <section ng-controller="HomeController">
-    <link rel="stylesheet" type="text/css" href="modules/core/client/css/splashpage1.css">
     <div class="splashpage">
         <div class="jumbotron text-center">
             <div class="row">
