+++ modules/projects/client/config/projects.client.routes.js
@@ -3,7 +3,7 @@
 // Setting up route
 angular.module('projects').config(['$stateProvider',
   function ($stateProvider) {
-    // Ratings state routing
+        // Ratings state routing
     $stateProvider
       .state('projects', {
         abstract: true,
@@ -14,20 +14,10 @@ angular.module('projects').config(['$stateProvider',
         url: '',
         templateUrl: 'modules/projects/client/views/list-project.client.view.html'
       })
-      .state('projects.create', {
-        url: '/create',
-        templateUrl: 'modules/projects/client/views/create-project.client.view.html',
-        data: {
-          roles: ['user', 'admin']
-        }
-      })
-      .state('projects.view', {
-        url: '/:projectId',
-        templateUrl: 'modules/projects/client/views/view-project.client.view.html'
-      })
-      .state('projects.edit', {
-        url: '/:projectId/edit',
-        templateUrl: 'modules/projects/client/views/edit-project.client.view.html',
+      .state('projects.rate', {
+        url: '/projects/rate/:projectId',
+        templateUrl: 'modules/ratings/client/views/single-rating.client.view.html',
+        controller: 'ProjectController',
         data: {
           roles: ['user', 'admin']
         }
