--- modules/users/client/services/admin/projects.client.service.js
@@ -1,14 +1,0 @@
-'use strict';
-
-// Projects service used for communicating with the ratings REST endpoints
-angular.module('users').factory('Projects', ['$resource',
-  function ($resource) {
-    return $resource('api/projects/:projectId', {
-      projectId: '@_id'
-    }, {
-      update: {
-        method: 'PUT'
-      }
-    });
-  }
-]);
