+++ modules/users/client/views/admin/edit-project.client.view.html
@@ -1,4 +1,4 @@
-<section ng-controller="AdminProjectController" ng-init="findOne()">
+<section ng-controller="ProjectController" ng-init="findOne()">
   <div class="container">
     <div class="col-md-10 col-md-offset-1">
       <div class="page-header">
