+++ modules/users/client/views/admin/list-project.client.view.html
@@ -1,4 +1,4 @@
-<section ng-controller="AdminProjectController" ng-init="find()">
+<section ng-controller="ProjectController" ng-init="find()">
     <div class="page-header">
         <div class="row">
             <div class="col-md-4">
