+++ modules/users/client/views/admin/view-project.client.view.html
@@ -1,4 +1,4 @@
-<section ng-controller="AdminProjectController" ng-init="findOne()">
+<section ng-controller="ProjectController" ng-init="findOne()">
     <div class="page-header">
         <h1 ng-bind="project.teamName"></h1>
     </div>
