--- modules/users/server/controllers/projects.server.controller.js
@@ -1,114 +1,0 @@
-'use strict';
-/* Need to update this to represent new schema */
-
-/**
- * Module dependencies.
- */
-var path = require('path'),
-  mongoose = require('mongoose'),
-  Project = mongoose.model('Project'),
-  errorHandler = require(path.resolve('./modules/core/server/controllers/errors.server.controller'));
-
-/**
- * Create a project
- */
-exports.create = function (req, res) {
-  var project = new Project(req.body);
-  project.teamName = req.body.teamName;
-  project.description = req.body.description;
-
-  project.save(function (err) {
-    if (err) {
-      return res.status(400).send({
-        message: errorHandler.getErrorMessage(err)
-      });
-    } else {
-      res.json(project);
-    }
-  });
-};
-
-/**
- * Show the current project
- */
-exports.read = function (req, res) {
-  res.json(req.project);
-};
-
-/**
- * Update a project
- */
-exports.update = function (req, res) {
-  var project = req.project;
-
-  project.teamName = req.body.teamName;
-  project.description = req.body.description;
-
-  console.log ("Server side: " + project.teamName + ", " + project.description);
-  project.save(function (err) {
-    if (err) {
-      return res.status(400).send({
-        message: errorHandler.getErrorMessage(err)
-      });
-    } else {
-      res.json(project);
-    }
-  });
-};
-
-/**
- * Delete an project
- */
-exports.delete = function (req, res) {
-  var project = req.project;
-
-  project.remove(function (err) {
-    if (err) {
-      return res.status(400).send({
-        message: errorHandler.getErrorMessage(err)
-      });
-    } else {
-      res.json(project);
-    }
-  });
-};
-
-/**
- * List of Projects
- */
-exports.list = function (req, res) {
-  Project.find().sort('-teamName').exec(function (err, projects) {
-    if (err) {
-      return res.status(400).send({
-        message: errorHandler.getErrorMessage(err)
-      });
-    } else {
-      res.json(projects);
-    }
-  });
-};
-
-/**
- * Project middleware
- */
-exports.projectByID = function (req, res, next, id) {
-
-  if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(400).send({
-      message: 'Project is invalid'
-    });
-  }
-
-  Project.findById(id).exec(function (err, project) {
-    if (err) {
-      return next(err);
-    } else if (!project) {
-      console.log('We couldnt find your project');
-      return res.status(404).send({
-        message: 'No project with that identifier has been found'
-      });
-    }
-    req.project = project;
-    next();
-  });
-};
