+++ modules/users/server/routes/users.server.routes.js
@@ -3,10 +3,11 @@
 module.exports = function (app) {
   // User Routes
   var users = require('../controllers/users.server.controller');
+  var admin = require('../controllers/admin.server.controller');
 
   // Setting up the users profile api
   app.route('/api/users/me').get(users.me);
-  app.route('/api/users').put(users.update);
+  app.route('/api/users').put(users.update).get(admin.listUser);
   app.route('/api/users/accounts').delete(users.removeOAuthProvider);
   app.route('/api/users/password').post(users.changePassword);
   app.route('/api/users/picture').post(users.changeProfilePicture);
